Solution for Day9

+++ AdventOfCode/Day5/Day5.cs
@@ -9,7 +9,7 @@ namespace AdventOfCode
         public static void Problem1(string input)
         {
             var lines = Misc.readLines(input, Environment.NewLine);
-            int[] values = new List<string>(lines[0].Split(",", StringSplitOptions.RemoveEmptyEntries)).ConvertAll((string val) => int.Parse(val)).ToArray();
+            long[] values = new List<string>(lines[0].Split(",", StringSplitOptions.RemoveEmptyEntries)).ConvertAll((string val) => long.Parse(val)).ToArray();
 
             var computer = new IntcodeComputer(values);
             computer.Run();
@@ -18,7 +18,7 @@ namespace AdventOfCode
         public static void Problem2(string input)
         {
             var lines = Misc.readLines(input, Environment.NewLine);
-            int[] values = new List<string>(lines[0].Split(",", StringSplitOptions.RemoveEmptyEntries)).ConvertAll((string val) => int.Parse(val)).ToArray();
+            long[] values = new List<string>(lines[0].Split(",", StringSplitOptions.RemoveEmptyEntries)).ConvertAll((string val) => long.Parse(val)).ToArray();
 
             var computer = new IntcodeComputer(values);
             computer.Run();
